move tabs-related functions to tabs.js
